data-generator/requirements.txt
kafka-python-ng==2.2.2
pandas==2.2.3
numpy==2.1.3
PyYAML==6.0.2
